backend/requirements.txt
fastapi
uvicorn
gspread
google-auth
requests
